move badge array into svelteui package
