+++ src/svelteui/data/BadgeArray.svelte
