+++ src/svelteui/data/BadgeArrayEmit.svelte
